Updated documentation.  Broke Oauth code into it own file since it does not make ajax requests.
