+++ src/main/webapp/resources/js/apis/galaxy/galaxy.js
@@ -1,55 +1,21 @@
 import axios from "axios";
 
+const GALAXY_AJAX_URL = `${window.TL.BASE_URL}ajax/galaxy-export`;
+
+/**
+ * Query the server to find out if the galaxy client has a valid authentication
+ * token.
+ * @param {string} clientId - the identifier for the galaxy client
+ * @returns {Promise<AxiosResponse<any> | never>}
+ */
 export const getGalaxyClientAuthentication = clientId =>
   axios
-    .get(
-      `${window.TL.BASE_URL}ajax/galaxy-export/authorized?clientId=${clientId}`
-    )
+    .get(`${GALAXY_AJAX_URL}/authorized?clientId=${clientId}`)
     .then(({ data }) => data);
 
+/**
+ * Get the samples that are in the cart in correct format to be sent to galaxy.
+ * @returns {Promise<AxiosResponse<any> | never>}
+ */
 export const getGalaxySamples = () =>
-  axios
-    .get(`${window.TL.BASE_URL}ajax/galaxy-export/samples`)
-    .then(({ data }) => data);
-
-export const exportToGalaxy = (
-  email,
-  makepairedcollection,
-  oauthCode,
-  oauthRedirect,
-  samples
-) => {
-  const name = `IRIDA-${Date.now()}`;
-
-  /*
-  This structure is expected by galaxy.
-   */
-  const params = {
-    _embedded: {
-      library: { name },
-      user: { email },
-      addtohistory: true, // Default according to Phil Mabon
-      makepairedcollection,
-      oauth2: {
-        code: oauthCode,
-        redirect: oauthRedirect
-      },
-      samples
-    }
-  };
-
-  /*
-  On the cart.html page, there is a form with the name "js-galaxy-form" that
-  get populated with the action (galaxy url).  It has a hidden input (#js-query)
-  that we set the stringified version of the galaxy parameters to.  After updating
-  the value, we submit the form, which will redirect us back to Galaxy.
-   */
-  const form = document.forms["js-galaxy-form"];
-  if (typeof form === "undefined") {
-    throw new Error(`Expecting the galaxy form with name "js-galaxy-form"`);
-  } else {
-    const input = form.elements["js-query"];
-    input.value = JSON.stringify(params);
-    form.submit();
-  }
-};
+  axios.get(`${GALAXY_AJAX_URL}/samples`).then(({ data }) => data);
